Debugging alsa_out

+++ latency_measurement/alsa_out.c
@@ -17,7 +17,7 @@ snd_pcm_format_t formatType;
 snd_pcm_access_t accessType;
 unsigned int channels;
 unsigned int samplingRate = 44100;
-unsigned char buffer[samplingRate*0.001];  /* some random data */
+unsigned char buffer[(int) samplingRate*0.001];  /* some random data */
 
 /* Display information about the PCM interface */
 void getHardwareParameters() {
